LayoutTests/fast/events/no-autoscroll-over-content-inset-expected.txt

PASS pageYOffset is pageYOffsetBefore
PASS successfullyParsed is true

TEST COMPLETE
Hello world
